requirements.txt
Babel==2.6.0
blinker==1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
coverage==4.5.1
dominate==2.3.4
factory-boy==2.11.1
Faker==0.8.16
Flask==1.0.2
Flask-BabelEx==0.9.3
Flask-Dance==1.0.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-nav==0.6
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
funcsigs==1.0.2
idna==2.7
ipaddress==1.0.22
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
loutilities==0.14.0rc2
lxml==4.2.5
MarkupSafe==1.0
mock==2.0.0
mysqlclient==1.3.13
nose2==0.7.4
oauthlib==2.1.0
passlib==1.7.1
pbr==4.0.4
python-dateutil==2.7.3
python-docx==0.8.7
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.9
SQLAlchemy-Utils==0.33.3
text-unidecode==1.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.23
URLObject==2.4.3
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
